Defined index types, logic for checking the validity of indices at compile-time.
Added `pmr` top-level module reexporting items for local macros.

Declared the `index_argument` module with the representation of indices passed to the
indexing macro.
The IndexArgument constructor also computes whether `IndexArgument`s are in order,
and which one is the maximum.

Declared the `index_properties` module which checks that the indices are disjoint when
they aren't ordered (if they are ordered they're disjoin by definition).

Declared the uniqueness module with an enum that represents whether there's overlapping
`IndexArgument`s in a slice.

Declared macros for:
- Constructing IndexArgumentsAndStats
- Constructing an arrays of `PrenormIndex`