Changed error handling to Result-based, defined an Error type.
Defined an Error type with all the errors in the crate,
and a method to panic at compile-time with the error message of the error.

Changed error handling in the whole crate to use `Result<T, Error>` or
an optional Error (alongside dummy data).

Renamed AreAllUnique to are AreAllDisjoint and added a generic payload to the No variant to
be able to put an Error in it.

Made `n ..= usize::MAX` indices an error,with an Error variant and custom error message.

Changed the indexing macro to:

    Use a new `ComputedConstants` struct instead of a `IndexArgumentsAndStats` constant,
    to allow the computed constant to add more fields without changing its name.

    Store an `Option<Error>` in the ComputedConstants,
    then checking for errors in a separate `const` initializer.
    Before doing this I got four separate errors when panicking in the
    constant that emited the error, now I only get one in the separate `const` initializer.

Made the `block` macro usable from the exported macros,
to bubble up errors without creating a new function.