Improved tests, added error for `usize::MAX` indices.
Added testing of invariants of IndexArguments,
used from both the example-based tests,and the automated testing.

In automated tests: made sure to test numbers close to `usize::MAX` as well as
close to the range of valid indices for the slice.

Added `UsizeMaxIndex__CurrentArgumentis` error when `usize::MAX` is passed as an index argument.